embedding-calculator/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==20.2.0
beautifulsoup4==4.12.2
cached-property==1.5.2
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colour==0.1.5
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.3
filelock==3.12.2
flasgger==0.9.5
Flask==1.1.2
fonttools==4.42.0
gast==0.3.3
gdown==4.5.4
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
h5py==2.10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.17.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.6.3
mccabe==0.7.0
mistune==3.0.1
networkx==3.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pip==23.0.1
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pydocstyle==6.3.0
pyflakes==3.1.0
pyparsing==3.1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==5.4.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==57.5.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==2.3.0
tf-slim==1.1.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tomli==2.0.1
tqdm==4.66.1
urllib3==2.0.4
Werkzeug==1.0.1
wheel==0.41.1
wrapt==1.15.0
zipp==3.16.2

# tests
mock==5.1.0
pytest==7.4.0
pytest-mock==3.11.1
requests==2.31.0
pylama==8.4.1

# dependencies for both scanner backends
Pillow==8.3.2
imagecodecs==2020.5.30
numpy==1.19.5
scipy==1.4.1
opencv-python==4.4.0.46
scikit-image==0.17.2
scikit-learn==0.23.2
joblib==0.17.0

# web server
uWSGI==2.0.19
